Move post-specific autosave logic to editor.js
