+++ packages/core-data/src/persist.js
@@ -1,4 +1,9 @@
 /**
+ * External dependencies
+ */
+import { identity } from 'lodash';
+
+/**
  * WordPress dependencies
  */
 import triggerFetch from '@wordpress/api-fetch';
@@ -18,7 +23,6 @@ export function persistRecordToAPI( {
 	__unstableFetch = triggerFetch,
 } ) {
 	return async ( { dispatch } ) => {
-		const recordId = getRecordId( entity, record );
 		let edits = record;
 		if ( entity.__unstablePrePersist ) {
 			edits = {
@@ -26,6 +30,7 @@ export function persistRecordToAPI( {
 				...entity.__unstablePrePersist( persistedRecord, edits ),
 			};
 		}
+		const recordId = getRecordId( entity, record );
 		const options = {
 			path: entityEndpointPath( entity, recordId ),
 			method: recordId ? 'PUT' : 'POST',
@@ -48,16 +53,11 @@ export function persistAutosaveToAPI( {
 	entity,
 	persistedRecord,
 	record,
+	prepareAutosaveRequest = identity,
+	reconcileAutosaveResponse = identity,
 	__unstableFetch = triggerFetch,
-	__prepareAutosaveRequest = prepareAutosaveRequest,
-	__reconcileAutosave = reconcileAutosave,
 } ) {
 	return async ( { select, dispatch } ) => {
-		const recordId = getRecordId( entity, record );
-		// Most of this autosave logic is very specific to posts.
-		// This is fine for now as it is the only supported autosave,
-		// but ideally this should all be handled in the back end,
-		// so the client just sends and receives objects.
 		const autosavePost = select.getAutosave(
 			persistedRecord.type,
 			persistedRecord.id,
@@ -68,12 +68,13 @@ export function persistAutosaveToAPI( {
 		// So we fallback to the previous autosave and then
 		// to the actual persisted entity if the edits don't
 		// have a value.
-		const requestData = __prepareAutosaveRequest( {
+		const requestData = prepareAutosaveRequest( {
 			...persistedRecord,
 			...autosavePost,
 			...record,
 		} );
 
+		const recordId = getRecordId( entity, record );
 		const path = entityEndpointPath( entity, recordId );
 		const updatedRecord = await __unstableFetch( {
 			path: `${ path }/autosaves`,
@@ -81,18 +82,14 @@ export function persistAutosaveToAPI( {
 			data: requestData,
 		} );
 
-		// An autosave may be processed by the server as a regular save
-		// when its update is requested by the author and the post had
-		// draft or auto-draft status.
+		// An autosave may be processed by the server as a regular save.
 		const processedAsRegularSave = persistedRecord.id === updatedRecord.id;
-		const reconciledRecord = __reconcileAutosave
-			? __reconcileAutosave(
-					persistedRecord,
-					requestData,
-					updatedRecord,
-					{ processedAsRegularSave }
-			  )
-			: updatedRecord;
+		const reconciledRecord = reconcileAutosaveResponse(
+			updatedRecord,
+			requestData,
+			persistedRecord,
+			{ processedAsRegularSave }
+		);
 
 		if ( processedAsRegularSave ) {
 			await dispatch.receiveEntityRecords(
@@ -111,38 +108,4 @@ export function persistAutosaveToAPI( {
 
 		return updatedRecord;
 	};
-}
-
-export function prepareAutosaveRequest( { title, excerpt, content, status } ) {
-	return {
-		title,
-		excerpt,
-		content,
-		status: status === 'auto-draft' ? 'draft' : status,
-	};
-}
-
-export function reconcileAutosave(
-	persistedRecord,
-	requestData,
-	updatedRecord,
-	{ processedAsRegularSave }
-) {
-	return processedAsRegularSave
-		? {
-				...persistedRecord,
-
-				title: updatedRecord.title,
-				excerpt: updatedRecord.excerpt,
-				content: updatedRecord.content,
-
-				// Status is only persisted in autosaves when going from
-				// "auto-draft" to "draft".
-				status:
-					persistedRecord.status === 'auto-draft' &&
-					requestData.status === 'draft'
-						? requestData.status
-						: persistedRecord.status,
-		  }
-		: updatedRecord;
 }
